delint
git-svn-id: https://svn.aros.org/svn/aros/trunk@53205 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/locale/countries/makecountry/countryprefs.c
@@ -24,17 +24,17 @@ unsigned long getCountryPrefsSize()
 
 unsigned long getCountryPrefsVers(struct IntCountryPrefs *cp)
 {
-    return cp->country_Version;
+    return (unsigned long)cp->country_Version;
 }
 
 unsigned long getCountryPrefsFlag(struct IntCountryPrefs *cp)
 {
-    return cp->country_Flag;
+    return (unsigned long)cp->country_Flag;
 }
 
 unsigned long getCountryNames(struct IntCountryPrefs *cp)
 {
-    return cp->country_NativeNames;
+    return (unsigned long)cp->country_NativeNames;
 }
 
 void convertEndianness(struct CountryPrefs *cp)
